Debug Alpha: implementing iTunes Film specification (version=5.2) . -separated asset processing logic from concrete metadata format (also generated beans) -separated metadata provider logic -embedded and implemented new lax iTunes Film schema (generated by official tool Transporter) -updated unit tests in accordance with modifications

+++ itunes-conversion/src/test/java/com/netflix/imfutility/itunes/util/ChaptersUtils.java
@@ -44,7 +44,9 @@ public final class ChaptersUtils {
 
     public static void createChapterFile(String chapterName) throws IOException {
         File file = new File(chapterName);
-        file.createNewFile();
+        if (!file.createNewFile()) {
+            throw new IOException(String.format("Cannot create test chapter file %s", file.getAbsolutePath()));
+        }
         file.deleteOnExit();
     }
 }
